# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nawa Real Estate Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  title: string;
  titleAr: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  descriptionAr?: string | null;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  order?: number;
  createdAt?: string;
}